fix: (21.05.21) Optimized counting unread messages on backend (#60)

+++ chat/src/main/kotlin/chatox/chat/api/response/ChatOfCurrentUserResponse.kt
@@ -11,7 +11,7 @@ data class ChatOfCurrentUserResponse(
         val name: String,
         val avatarUri: String?,
         val chatParticipation: ChatParticipationMinifiedResponse,
-        val unreadMessagesCount: Int,
+        val unreadMessagesCount: Long,
         val lastMessage: MessageResponse?,
         val createdAt: ZonedDateTime,
         val description: String?,
